requirements.txt
# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
llama-index==0.9.30

# Vector Databases (choose one or use multiple)
qdrant-client==1.7.0
chromadb==0.4.18
pinecone-client==2.2.4
weaviate-client==3.25.3

# Document Processing
unstructured[all-docs]==0.11.6
pypdf==3.17.4
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
beautifulsoup4==4.12.2
tiktoken==0.5.2

# Embedding Models
sentence-transformers==2.2.2
openai==1.6.1
cohere==4.37

# Advanced Features
ragas==0.0.22  # RAG evaluation
instructor==0.4.6  # Structured outputs
pydantic==2.5.2
asyncio==3.4.3

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
tqdm==4.66.1
rich==13.7.0

# Monitoring & Observability
langsmith==0.0.69
wandb==0.16.1